Don’t assume .well-known must be a symbolic link.
The .well-known directory has multiple uses, however by forcing it to
be a symbolic link to a directory controlled by this script, the script
prevents user from using the directory for purposes other than ACME
authorisation (such as Web Key Directory).

Change tls-setup.sh so that when .well-known directory already exists
in site root, it’ll create acme-challenge symbolic link inside of it
thus playing nicely with other uses of the directory.